Fix kafka issues

+++ src/main/java/com/auction/product/dto/ProductPurchaseRequest.java
@@ -6,6 +6,6 @@ public record ProductPurchaseRequest(
         @NotNull(message = "Product Id mandatory")
         Long productId,
         @NotNull(message = "Quantity Id mandatory")
-        double quantity
+        int quantity
 ) {
 }
